Increasing code maintainability: Fix 1
