+++ server/routes/auth/auth.js
@@ -1,0 +1,11 @@
+const express = require("express");
+const {
+  registerUser,
+  loginUser,
+} = require("../../controllers/auth/authController.js");
+const router = express.Router();
+
+router.post("/register", registerUser);
+router.post("/login", loginUser);
+
+module.exports = router;
