+++ server/utils/index.js
@@ -1,4 +1,5 @@
 const connectDB = require("./db/connectDB");
 const corsConfig = require("./cors/corsConfig");
+const { hashPassword, comparePassword } = require("./bcrypt/bcryptUtils");
 
-module.exports = { connectDB, corsConfig };
+module.exports = { connectDB, corsConfig, hashPassword, comparePassword };
